Now multiplies speed by frametime

+++ src/game/ball.h
@@ -8,9 +8,9 @@ class Ball
         void Draw();
 
     private:
-        int x;
-        int y;
-        int speedX;
-        int speedY;
-        int radius;
+        float x;
+        float y;
+        float speedX;
+        float speedY;
+        float radius;
 };
